langchain_basic/requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
python-dotenv
# langchain_community
langchain_ollama
langserve
